chore(hybridgateway): generalize l4 route conflict logics
